% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addDayToDate}
\alias{addDayToDate}
\title{Add day to date string}
\usage{
addDayToDate(x, day = "01")
}
\arguments{
\item{x}{\code{character} date.}

\item{day}{\code{character} to append.}
}
\value{
\code{character} with day added.
}
\description{
Add day to date string
}
