% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{appendAccountingEntity}
\alias{appendAccountingEntity}
\title{Append accouniting entity to database}
\usage{
appendAccountingEntity(accEntity)
}
\arguments{
\item{accEntity}{\code{data.frame} with accounting entity info.}
}
\value{
\code{NULL}
}
\description{
Append accouniting entity to database. Also appends data to junction table.
}
