% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{appendFinancialReport}
\alias{appendFinancialReport}
\title{Append financial report to database}
\usage{
appendFinancialReport(finReport)
}
\arguments{
\item{finReport}{\code{data.frame} with financial statement info.}
}
\value{
\code{NULL}
}
\description{
Append financial report to database. Also appends data to junction table.
}
