% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{appendFinancialStatement}
\alias{appendFinancialStatement}
\title{Append financial statement to database}
\usage{
appendFinancialStatement(finStatement)
}
\arguments{
\item{finStatement}{\code{data.frame} with financial statement info.}
}
\value{
\code{NULL}
}
\description{
Append financial statement to database. Also appends data to junction table.
}
