% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-utils.R
\name{appendIfMissing}
\alias{appendIfMissing}
\title{Append data if missing}
\usage{
appendIfMissing(x, tableName, con)
}
\arguments{
\item{x}{\code{data.frame} to append.}

\item{tableName}{\code{character} name of the table where to append.}

\item{con}{database connection}
}
\value{
\code{NULL}
}
\description{
Append missing data to database.
}
