% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-utils.R
\name{createUrl}
\alias{createUrl}
\title{Create query}
\usage{
createUrl(endpoint, ..., baseUrl = "http://www.registeruz.sk/cruz-public/api")
}
\arguments{
\item{endpoint}{`character` endpoint.}

\item{...}{params that should be sent to endpoint}

\item{baseUrl}{`character` base url of the API.}
}
\value{
`character` url.
}
\description{
Creates query from the endpoints and parameters.
}
