% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{getAccountingEntityDetails}
\alias{getAccountingEntityDetails}
\title{Get details of accounting entity}
\usage{
getAccountingEntityDetails(id)
}
\arguments{
\item{id}{\code{numeric} id of the entity.}
}
\value{
\code{list} with details.
}
\description{
Get details of accounting entity.
}
