% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllChangedIDs.R
\name{getAllChangedIDs}
\alias{getAllChangedIDs}
\alias{getAllChangedFinancialReports}
\alias{getAllChangedFinancialStatements}
\alias{getAllChangedAccountingEntities}
\title{Get all changed IDs since date}
\usage{
getAllChangedIDs(from, maxNumber = 1000, fun)

getAllChangedFinancialReports(from, maxNumber = 1000)

getAllChangedFinancialStatements(from, maxNumber = 1000)

getAllChangedAccountingEntities(from, maxNumber = 1000)
}
\arguments{
\item{from}{\code{character} date}

\item{maxNumber}{\code{numeric} max number which should be downloaded in one batch.}

\item{fun}{\code{function} for getting IDs.}
}
\value{
\code{numeric} \code{vector} with changed IDs
}
\description{
Get all changed IDs since date.
}
\section{Functions}{
\itemize{
\item \code{getAllChangedFinancialReports}: Get all changed financial reports.

\item \code{getAllChangedFinancialStatements}: Get all changed financial statements.

\item \code{getAllChangedAccountingEntities}: Get all changed accounting entities.
}}

