% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{getBaseFinReport}
\alias{getBaseFinReport}
\title{Get base of financial report}
\usage{
getBaseFinReport(finReport)
}
\arguments{
\item{finReport}{financial report \code{list}.}
}
\value{
\code{data.frame} with base of financial report.
}
\description{
Get base of financial report.
}
