% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-utils.R
\name{getDBConnection}
\alias{getDBConnection}
\title{Get DB connection}
\usage{
getDBConnection()
}
\value{
DB connection
}
\description{
Returns existing DB connection, if the connection does not exist yet, it is created.
}
