% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{getFinancialReportDetails}
\alias{getFinancialReportDetails}
\title{Get details of financial report}
\usage{
getFinancialReportDetails(id)
}
\arguments{
\item{id}{\code{numeric} id of the report.}
}
\value{
\code{list} with details.
}
\description{
Get details of financial report.
}
