% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{getFinancialStatementDetails}
\alias{getFinancialStatementDetails}
\title{Get details of financial statement}
\usage{
getFinancialStatementDetails(id)
}
\arguments{
\item{id}{\code{numeric} id of the statement.}
}
\value{
\code{list} with details.
}
\description{
Get details of financial statement.
}
