% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{getNumericReportData}
\alias{getNumericReportData}
\alias{getAktivaFinReport}
\alias{getPasivaFinReport}
\alias{getZSFinReport}
\title{Get data from financial report on position}
\usage{
getNumericReportData(finReport, position)

getAktivaFinReport(finReport)

getPasivaFinReport(finReport)

getZSFinReport(finReport)
}
\arguments{
\item{finReport}{financial report \code{list}.}

\item{position}{\code{numeric} position of data which should be selected.}
}
\value{
\code{data.frame} with the subsetted data.
}
\description{
Get data from financial report on position.
}
\section{Functions}{
\itemize{
\item \code{getAktivaFinReport}: Get assets data from financial report.

\item \code{getPasivaFinReport}: Get liabilities and equity data from financial report.

\item \code{getZSFinReport}: Get income statement data from financial report.
}}

