% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-utils.R
\name{getObjectFromURL}
\alias{getObjectFromURL}
\title{Return list object from URL}
\usage{
getObjectFromURL(url)
}
\arguments{
\item{url}{\code{character} from where the data should be read.}
}
\value{
\code{list} with the read data.
}
\description{
Read lines from the URL and returns results as \code{list}.
}
