% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{getTitleFinReport}
\alias{getTitleFinReport}
\title{Get title page of financial report}
\usage{
getTitleFinReport(finReport)
}
\arguments{
\item{finReport}{financial report \code{list}.}
}
\value{
\code{data.frame} with title page of financial report.
}
\description{
Get title page of financial report.
}
