% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toCharDF}
\alias{toCharDF}
\title{Transform all columns to character}
\usage{
toCharDF(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{chracater} \code{data.frame}.
}
\description{
Transform all columns to character.
}
