% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-utils.R
\name{tryUntilSuccess}
\alias{tryUntilSuccess}
\title{Query URL multiple times}
\usage{
tryUntilSuccess(url, numberOfTries = 20, fun = getObjectFromURL)
}
\arguments{
\item{url}{\code{character} url to query.}

\item{numberOfTries}{\code{numeric} number of tries that should be done.}

\item{fun}{\code{function} used for quering.}
}
\value{
\code{object}
}
\description{
Queries URL 'numberOfTries' until success or warning for url is logged.
}
